enhance any_sender_of to permit type-erased receiver queries

+++ test/bulk_schedule_test.cpp
@@ -103,7 +103,7 @@ TEST(bulk, Pipeable) {
         }, unifex::par_unseq)
       | unifex::bulk_transform(
         [&](std::size_t index) noexcept {
-            output[index] = index;
+            output[index] = static_cast<int>(index);
         }, unifex::par_unseq)
       | unifex::bulk_join()
       | unifex::sync_wait();
